Update styling on auth screens. Add menu bar icons

+++ Pulse/AppDelegate.h
@@ -38,7 +38,7 @@
 +(BOOL)validateEmail:(NSString *)email ;
 +(BOOL)isValidCharacter:(NSString*)string filterCharSet:(NSString*)set;
 -(NSString*)formatNumber:(NSString*)mobileNumber;
--(int)getLength:(NSString*)mobileNumber;
+-(NSInteger)getLength:(NSString*)mobileNumber;
 
 -(void)userLogout;
